+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Removed access token validation (Issue 4)
 * Allowed access token response to continue when no refresh token provided (Issue 6)
+* Various dictionary value lookups improved to account for missing keys (Issue 7)
 
 ## Released