+++ CHANGES.md
@@ -2,10 +2,19 @@
 
 ## Unreleased
 
-* Removed access token validation (Issue 4)
-* Allowed access token response to continue when no refresh token provided (Issue 6)
+
+
 
 ## Released
+
+### Version 0.0.1a10
+
+Released 2021-10-21
+
+* Major improvements to the cli, now uses argparse, docs updated
+* Removed access token validation (Issue 4)
+* Allowed access token response to continue when no refresh token provided (Issue 6)
+* Various dictionary value lookups improved to account for missing keys (Issue 7)
 
 ### Version 0.0.1a9